Prometheus exporter works, but ...
This "works" now. It's nicely pushing a counter and a gauge for every
cmd/step/build. But the prometheus aggregation gateway can't aggregate
gauges in a nice way. At the moment it just calculates a sum. Is that
useful? Not sure. Metrics build this kind of stuff seem super hard
without proper push support.